requirements.txt
numpy
matplotlib
scipy
Pillow
opencv-python